ai-service/requirements.txt
flask==3.0.0
torch==2.5.0
torchvision==0.20.0
pillow==10.1.0
numpy<2.0
